HwAccel Multi-CU: update to 2021.2(#185)
- Removed OpenCV dependency
- Other minor changes
